+++ src/main/java/project/forwork/api/domain/resume/controller/ResumeAdminController.java
@@ -10,13 +10,12 @@ import project.forwork.api.common.annotation.Current;
 import project.forwork.api.common.api.Api;
 import project.forwork.api.domain.resume.controller.model.ResumeDetailResponse;
 import project.forwork.api.domain.resume.controller.model.ResumePage;
-import project.forwork.api.domain.resume.infrastructure.ResumeSearchCond;
+import project.forwork.api.domain.resume.infrastructure.enums.PageStep;
 import project.forwork.api.domain.resume.infrastructure.enums.PeriodCond;
 import project.forwork.api.domain.resume.infrastructure.enums.ResumeStatus;
 import project.forwork.api.domain.resume.model.Resume;
 import project.forwork.api.domain.resume.service.ResumeService;
 import project.forwork.api.common.domain.CurrentUser;
-
 import java.time.LocalDateTime;
 
 @RestController
@@ -39,7 +38,22 @@ public class ResumeAdminController {
         return Api.OK(ResumeDetailResponse.from(resume));
     }
 
-    @Operation(summary = "첫 페이지 resume 요청 조회 API", description = "필터링 및 정렬 조건을 선택 할 경우 첫 페이지로 이동")
+    @GetMapping("/page")
+    public Api<ResumePage> getFilteredPage(
+            @RequestParam(required = false) PeriodCond periodCond,
+            @RequestParam(required = false) ResumeStatus status,
+            @RequestParam(defaultValue = "FIRST") PageStep pageStep,
+            @RequestParam(required = false) LocalDateTime lastModifiedAt,
+            @RequestParam(required = false) Long lastId,
+            @RequestParam(defaultValue = "6") int limit
+    ) {
+        // 필터링 및 페이징을 처리하는 서비스 호출
+        ResumePage result = resumeService
+                .getFilteredAndPagedResults(periodCond, status, pageStep, lastModifiedAt, lastId, limit);
+        return Api.OK(result);
+    }
+
+/*    @Operation(summary = "첫 페이지 resume 요청 조회 API", description = "필터링 및 정렬 조건을 선택 할 경우 첫 페이지로 이동")
     @GetMapping("/first")
     public Api<ResumePage> findFirstPage(
             @RequestParam(required = false) PeriodCond periodCond,
@@ -85,5 +99,5 @@ public class ResumeAdminController {
     ){
         ResumePage resumePage = resumeService.findPreviousPage(periodCond, status, lastModifiedAt, lastId, limit);
         return Api.OK(resumePage);
-    }
+    }*/
 }
