+++ src/test/java/project/forwork/api/domain/resumedecision/service/ResumeDecisionServiceTest.java
@@ -3,8 +3,8 @@ package project.forwork.api.domain.resumedecision.service;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import project.forwork.api.common.domain.CurrentUser;
-import project.forwork.api.domain.resume.infrastructure.enums.FieldType;
-import project.forwork.api.domain.resume.infrastructure.enums.LevelType;
+import project.forwork.api.common.infrastructure.enums.FieldType;
+import project.forwork.api.common.infrastructure.enums.LevelType;
 import project.forwork.api.domain.resume.infrastructure.enums.ResumeStatus;
 import project.forwork.api.domain.resume.model.Resume;
 import project.forwork.api.domain.user.infrastructure.enums.RoleType;
